Update to improve compatibility with new programming cable
Fixed bug in win32 open com port, ports above COM9 were not handled correctly.

Improved handling of initialization, the code is now even more specific to the cortex.

Changed the read/write transfer status display to be more friendly to the eclipse IDE.  Eclipse does not handle "/r" so status was scrolling of the console.

Signed-off-by: James Pearman <jbpearman@mac.com>